requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.1
selenium==4.15.2
webdriver-manager==4.0.1
fake-useragent==1.4.0
python-dotenv==1.0.0
gunicorn==21.2.0 